src: fix usage of `napi_extended_error_info` in `Error::New()`
All fields of the `napi_extended_error_info` structure gets reset in
subsequent Node-API function calls on the same `env`. This includes a
call to `napi_is_exception_pending()`. So here it is necessary to make
a copy of the information as the `error_code` field is used later on.

Signed-off-by: Darshan Sen <darshan.sen@postman.com>

PR-URL: https://github.com/nodejs/node-addon-api/pull/1092
Fixes: https://github.com/nodejs/node-addon-api/issues/1089
Reviewed-By: Michael Dawson <michael_dawson@ca.ibm.com>